chore: simplify removal of circ references
